pyproject.toml
[project]
name = "flickfeast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "google-auth>=2.30.0",
    "openai",
    "openai-agents",
    "pydantic>=2.7.0",
    "python-dotenv>=1.2.1",
    "pytest>=8.2.0",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.30.0",
]
